+++ lib/Frontend/CMakeLists.txt
@@ -13,10 +13,12 @@ add_swift_host_library(swiftFrontend STATIC
   ModuleInterfaceLoader.cpp
   ModuleInterfaceSupport.cpp
   PrintingDiagnosticConsumer.cpp
+  Serialization.cpp
   SerializedDiagnosticConsumer.cpp)
 add_dependencies(swiftFrontend
   SwiftOptions)
 target_link_libraries(swiftFrontend PRIVATE
+  swiftAPIDigester
   swiftAST
   swiftConstExtract
   swiftIDE
@@ -28,6 +30,6 @@ target_link_libraries(swiftFrontend PRIVATE
   swiftLocalization
   swiftSema
   swiftSerialization
-  swiftAPIDigester)
+  swiftSymbolGraphGen)
 
 set_swift_llvm_is_available(swiftFrontend)
