+++ tools/swift-dependency-tool/CMakeLists.txt
@@ -7,5 +7,6 @@ target_link_libraries(swift-dependency-tool
                       PRIVATE
                         swiftAST
                         swiftParse
+                        swiftSerialization
                         swiftClangImporter)
